Continue the refactor
