+++ Cognite.Simulator.Utils/StateUtils.cs
@@ -1,4 +1,5 @@
-﻿using Cognite.Extractor.StateStorage;
+﻿using Cognite.Extensions;
+using Cognite.Extractor.StateStorage;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -23,28 +24,32 @@ namespace Cognite.Simulator.Utils
             var stateToDelete = col
                 .Find(s => !idsToKeep.Contains(s.Id))
                 .ToList();
-            if (stateToDelete.Any())
+            var filesInUseMap = col
+                    .Find(f => !string.IsNullOrEmpty(f.FilePath) && idsToKeep.Contains(f.Id))
+                    .ToDictionarySafe(f => f.FilePath, f => true);
+
+            foreach (var state in stateToDelete)
             {
-                foreach(var state in stateToDelete)
+                if (!filesInUseMap.ContainsKey(state.FilePath))
                 {
-                    if (state.IsInDirectory) {
-                        // get directory path from file path 
-                        // (file path is the directory path + file name)
+                    if (state.IsInDirectory)
+                    {
                         var dirPath = Path.GetDirectoryName(state.FilePath);
                         DeleteLocalDirectory(dirPath);
                     } else {
                         DeleteLocalFile(state.FilePath);
                     }
-                    col.Delete(state.Id);
                 }
+                col.Delete(state.Id);
             }
         }
 
-        internal static async Task RemoveFileStates(
+        internal static async Task RemoveFileStates<FileStateType>(
             this IExtractionStateStore store,
             string tableName,
-            IEnumerable<(FileState, bool)> states,
+            IEnumerable<(FileStateType, bool)> states,
             CancellationToken token)
+            where FileStateType : FileState
         {
             if (!states.Any())
             {
